requirements.txt
accelerate>=0.33
diffusers>=0.31
imageio>=2.34
imageio-ffmpeg>=0.5
numpy>=1.24
pillow>=10.0
protobuf>=4.25
pypdf>=4.2
safetensors>=0.4
sentencepiece>=0.2
tiktoken>=0.7
torch>=2.2
transformers>=4.41
